+++ diktat-common/src/main/kotlin/org/cqfn/diktat/common/config/reader/JsonResourceConfigReader.kt
@@ -2,23 +2,25 @@ package org.cqfn.diktat.common.config.reader
 
 import mu.KotlinLogging
 
+import java.io.BufferedReader
 import java.io.IOException
-import java.io.InputStream
 
 /**
  * This class is used to read some resource in any format that you will specify.
  * Usage:
  * 1) implement this class with implementing two methods:
  *    a. getConfigFile - to return URI (path) to the file that you want to read
- *    b. parse - implement parser for your file format (for example parse it to a proper json)
+ *    b. parseResource  - implement parser for your file format (for example parse it to a proper json)
  * 2) Use your new class MyReader(javaClass.classLoader).readResource("some/path/to/file.format")
  *
- * @param classLoader The [ClassLoader] used to load the requested resource.
  * @param T - class name parameter that will be used in calculation of classpath
  */
-abstract class AbstractResourceConfigReader<T : Any>(
-    protected val classLoader: ClassLoader
-) : AbstractConfigReader<T>() {
+abstract class JsonResourceConfigReader<T> {
+    /**
+     * The [ClassLoader] used to load the requested resource.
+     */
+    abstract val classLoader: ClassLoader
+
     /**
      * @param resourceFileName - related path to a file from resources
      * @return object of type [T] if resource has been parsed successfully
@@ -27,7 +29,7 @@ abstract class AbstractResourceConfigReader<T : Any>(
         val resourceStream = getConfigFile(resourceFileName)
         resourceStream?.let {
             try {
-                return parse(it)
+                return parseResource(it)
             } catch (e: IOException) {
                 log.error("Cannot read config file $resourceFileName due to: ", e)
             }
@@ -40,9 +42,17 @@ abstract class AbstractResourceConfigReader<T : Any>(
      * you can override this method in case you would like to read a file not simply from resources
      *
      * @param resourceFileName name of the resource which will be loaded using [classLoader]
-     * @return [InputStream] representing loaded resource
+     * @return [BufferedReader] representing loaded resource
      */
-    protected open fun getConfigFile(resourceFileName: String): InputStream? = classLoader.getResourceAsStream(resourceFileName)
+    protected open fun getConfigFile(resourceFileName: String): BufferedReader? =
+        classLoader.getResourceAsStream(resourceFileName)?.bufferedReader()
+    /**
+     * you can specify your own parser, in example for parsing stream as a json
+     *
+     * @param fileStream a [BufferedReader] representing loaded resource file
+     * @return resource parsed as type [T]
+     */
+    protected abstract fun parseResource(fileStream: BufferedReader): T
 
     companion object {
         private val log = KotlinLogging.logger {}
