+++ diktat-rules/src/test/kotlin/org/cqfn/diktat/ruleset/smoke/RulesConfigValidationTest.kt
@@ -1,6 +1,5 @@
 package org.cqfn.diktat.ruleset.smoke
 
-import org.cqfn.diktat.ruleset.rules.DiktatRuleConfigReaderImpl
 import org.cqfn.diktat.ruleset.rules.DiktatRuleSetFactoryImpl
 import org.cqfn.diktat.test.framework.util.deleteIfExistsSilently
