+++ diktat-test-framework/src/main/kotlin/org/cqfn/diktat/test/framework/config/TestConfigReader.kt
@@ -1,6 +1,6 @@
 package org.cqfn.diktat.test.framework.config
 
-import org.cqfn.diktat.common.config.reader.AbstractResourceConfigReader
+import org.cqfn.diktat.common.config.reader.JsonResourceConfigReader
 
 import java.io.IOException
 import java.io.InputStream
@@ -10,10 +10,10 @@ import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.decodeFromStream
 
 /**
- * A [AbstractResourceConfigReader] to read tests configuration as [TestConfig]
+ * A [JsonResourceConfigReader] to read tests configuration as [TestConfig]
  * @property classLoader a [ClassLoader] to load configuration file
  */
-class TestConfigReader(configFilePath: String, classLoader: ClassLoader) : AbstractResourceConfigReader<TestConfig>(classLoader) {
+class TestConfigReader(configFilePath: String, classLoader: ClassLoader) : JsonResourceConfigReader<TestConfig>(classLoader) {
     /**
      * The [TestConfig] which is read from
      */
